[core] Initialize states before feature hooks
