+++ packages/grid/_modules_/grid/hooks/features/columnPinning/useGridColumnPinning.tsx
@@ -20,14 +20,34 @@ import {
   GridPinnedPosition,
 } from '../../../models/api/gridColumnPinningApi';
 import { gridPinnedColumnsSelector } from './columnPinningSelector';
-import { useGridStateInit } from '../../utils/useGridStateInit';
 import { useGridSelector } from '../../utils/useGridSelector';
 import { filterColumns } from '../../../../../x-data-grid-pro/src/DataGridProVirtualScroller';
 import { GridRowParams } from '../../../models/params/gridRowParams';
 import { MuiEvent } from '../../../models/muiEvent';
 import { GridState } from '../../../models';
+import { GridStateInitializer } from '../../utils/useGridInitializeState';
 
 const Divider = () => <MuiDivider onClick={(event) => event.stopPropagation()} />;
+
+export const columnPinningStateInitializer: GridStateInitializer<
+  Pick<DataGridProProcessedProps, 'pinnedColumns' | 'initialState' | 'disableColumnPinning'>
+> = (state, props) => {
+  let model: GridPinnedColumns;
+  if (props.disableColumnPinning) {
+    model = {};
+  } else if (props.pinnedColumns) {
+    model = props.pinnedColumns;
+  } else if (props.initialState?.pinnedColumns) {
+    model = props.initialState?.pinnedColumns;
+  } else {
+    model = {};
+  }
+
+  return {
+    ...state,
+    pinnedColumns: model,
+  };
+};
 
 const mergeStateWithPinnedColumns =
   (pinnedColumns: GridPinnedColumns) =>
@@ -40,23 +60,6 @@ export const useGridColumnPinning = (
     'initialState' | 'disableColumnPinning' | 'pinnedColumns' | 'onPinnedColumnsChange'
   >,
 ): void => {
-  useGridStateInit(apiRef, (state) => {
-    let model: GridPinnedColumns;
-    if (props.disableColumnPinning) {
-      model = {};
-    } else if (props.pinnedColumns) {
-      model = props.pinnedColumns;
-    } else if (props.initialState?.pinnedColumns) {
-      model = props.initialState?.pinnedColumns;
-    } else {
-      model = {};
-    }
-
-    return {
-      ...state,
-      pinnedColumns: model,
-    };
-  });
   const pinnedColumns = useGridSelector(apiRef, gridPinnedColumnsSelector);
 
   // Each visible row (not to be confused with a filter result) is composed of a central .MuiDataGrid-row element
@@ -175,33 +178,6 @@ export const useGridColumnPinning = (
     [props.disableColumnPinning],
   );
 
-  const reorderPinnedColumns = React.useCallback<GridPreProcessor<'hydrateColumns'>>(
-    (columnsState) => {
-      if (columnsState.all.length === 0 || props.disableColumnPinning) {
-        return columnsState;
-      }
-
-      const [leftPinnedColumns, rightPinnedColumns] = filterColumns(
-        pinnedColumns,
-        columnsState.all,
-      );
-
-      if (leftPinnedColumns.length === 0 && rightPinnedColumns.length === 0) {
-        return columnsState;
-      }
-
-      const centerColumns = columnsState.all.filter((field) => {
-        return !leftPinnedColumns.includes(field) && !rightPinnedColumns.includes(field);
-      });
-
-      return {
-        ...columnsState,
-        all: [...leftPinnedColumns, ...centerColumns, ...rightPinnedColumns],
-      };
-    },
-    [pinnedColumns, props.disableColumnPinning],
-  );
-
   const checkIfCanBeReordered = React.useCallback<GridPreProcessor<'canBeReordered'>>(
     (initialValue, { targetIndex }) => {
       const visibleColumnFields = gridVisibleColumnFieldsSelector(apiRef);
@@ -261,7 +237,6 @@ export const useGridColumnPinning = (
 
   useGridRegisterPreProcessor(apiRef, 'scrollToIndexes', calculateScrollLeft);
   useGridRegisterPreProcessor(apiRef, 'columnMenu', addColumnMenuButtons);
-  useGridRegisterPreProcessor(apiRef, 'hydrateColumns', reorderPinnedColumns);
   useGridRegisterPreProcessor(apiRef, 'canBeReordered', checkIfCanBeReordered);
   useGridRegisterPreProcessor(apiRef, 'exportState', stateExportPreProcessing);
   useGridRegisterPreProcessor(apiRef, 'restoreState', stateRestorePreProcessing);
