+++ packages/grid/_modules_/grid/hooks/features/detailPanel/useGridDetailPanel.ts
@@ -2,10 +2,7 @@ import * as React from 'react';
 import { GridApiRef } from '../../../models/api/gridApiRef';
 import { useGridRegisterPreProcessor } from '../../core/preProcessing';
 import { useGridStateInit } from '../../utils/useGridStateInit';
-import {
-  GRID_DETAIL_PANEL_TOGGLE_COL_DEF,
-  GRID_DETAIL_PANEL_TOGGLE_FIELD,
-} from './gridDetailPanelToggleColDef';
+import { GRID_DETAIL_PANEL_TOGGLE_FIELD } from './gridDetailPanelToggleColDef';
 import { useGridApiEventHandler } from '../../utils/useGridApiEventHandler';
 import { GridEvents } from '../../../models/events/gridEvents';
 import { GridEventListener } from '../../../models/events/gridEventListener';
@@ -82,33 +79,6 @@ export const useGridDetailPanel = (
   useGridApiEventHandler(apiRef, GridEvents.cellClick, handleCellClick);
   useGridApiEventHandler(apiRef, GridEvents.cellKeyDown, handleCellKeyDown);
 
-  const addToggleColumn = React.useCallback<GridPreProcessor<'hydrateColumns'>>(
-    (columnsState) => {
-      if (props.getDetailPanelContent == null) {
-        // Remove the toggle column, when it exists
-        if (columnsState.lookup[GRID_DETAIL_PANEL_TOGGLE_FIELD]) {
-          delete columnsState.lookup[GRID_DETAIL_PANEL_TOGGLE_FIELD];
-          columnsState.all = columnsState.all.filter(
-            (field) => field !== GRID_DETAIL_PANEL_TOGGLE_FIELD,
-          );
-        }
-        return columnsState;
-      }
-
-      // Don't add the toggle column if there's already one
-      // The user might have manually added it to have it in a custom position
-      if (columnsState.lookup[GRID_DETAIL_PANEL_TOGGLE_FIELD]) {
-        return columnsState;
-      }
-
-      // Othewise, add the toggle column at the beginning
-      columnsState.all = [GRID_DETAIL_PANEL_TOGGLE_FIELD, ...columnsState.all];
-      columnsState.lookup[GRID_DETAIL_PANEL_TOGGLE_FIELD] = GRID_DETAIL_PANEL_TOGGLE_COL_DEF;
-      return columnsState;
-    },
-    [props.getDetailPanelContent],
-  );
-
   const addDetailHeight = React.useCallback<GridPreProcessor<'rowHeight'>>(
     (initialValue, row) => {
       if (expandedRowIds.length === 0 || !expandedRowIds.includes(row.id)) {
@@ -123,7 +93,6 @@ export const useGridDetailPanel = (
     [apiRef, expandedRowIds],
   );
 
-  useGridRegisterPreProcessor(apiRef, 'hydrateColumns', addToggleColumn);
   useGridRegisterPreProcessor(apiRef, 'rowHeight', addDetailHeight);
 
   apiRef.current.unstable_updateControlState({
