+++ packages/grid/_modules_/grid/hooks/features/pagination/useGridPageSize.ts
@@ -9,10 +9,32 @@ import {
   useGridApiEventHandler,
   useGridSelector,
 } from '../../utils';
-import { useGridStateInit } from '../../utils/useGridStateInit';
 import { gridPageSizeSelector } from './gridPaginationSelector';
 import { gridDensityRowHeightSelector } from '../density';
 import { GridPreProcessor, useGridRegisterPreProcessor } from '../../core/preProcessing';
+import { GridStateInitializer } from '../../utils/useGridInitializeState';
+
+const defaultPageSize = (autoPageSize: boolean) => (autoPageSize ? 0 : 100);
+
+export const pageSizeStateInitializer: GridStateInitializer<
+  Pick<DataGridProcessedProps, 'pageSize' | 'initialState' | 'autoPageSize'>
+> = (state, props) => {
+  let pageSize: number;
+  if (props.pageSize != null) {
+    pageSize = props.pageSize;
+  } else if (props.initialState?.pagination?.pageSize != null) {
+    pageSize = props.initialState.pagination.pageSize;
+  } else {
+    pageSize = defaultPageSize(props.autoPageSize);
+  }
+
+  return {
+    ...state,
+    pagination: {
+      pageSize,
+    },
+  };
+};
 
 const mergeStateWithPageSize =
   (pageSize: number) =>
@@ -26,7 +48,6 @@ const mergeStateWithPageSize =
 
 /**
  * @requires useGridDimensions (event) - can be after
- * @requires useGridFilter (state)
  */
 export const useGridPageSize = (
   apiRef: GridApiRef,
@@ -37,26 +58,6 @@ export const useGridPageSize = (
 ) => {
   const logger = useGridLogger(apiRef, 'useGridPageSize');
   const rowHeight = useGridSelector(apiRef, gridDensityRowHeightSelector);
-
-  const defaultPageSize = props.autoPageSize ? 0 : 100;
-
-  useGridStateInit(apiRef, (state) => {
-    let pageSize: number;
-    if (props.pageSize != null) {
-      pageSize = props.pageSize;
-    } else if (props.initialState?.pagination?.pageSize != null) {
-      pageSize = props.initialState.pagination.pageSize;
-    } else {
-      pageSize = defaultPageSize;
-    }
-
-    return {
-      ...state,
-      pagination: {
-        pageSize,
-      },
-    };
-  });
 
   apiRef.current.unstable_updateControlState({
     stateId: 'pageSize',
@@ -95,7 +96,7 @@ export const useGridPageSize = (
   const stateExportPreProcessing = React.useCallback<GridPreProcessor<'exportState'>>(
     (prevState) => {
       const pageSizeToExport = gridPageSizeSelector(apiRef);
-      if (pageSizeToExport === defaultPageSize) {
+      if (pageSizeToExport === defaultPageSize(props.autoPageSize)) {
         return prevState;
       }
 
@@ -107,7 +108,7 @@ export const useGridPageSize = (
         },
       };
     },
-    [apiRef, defaultPageSize],
+    [apiRef, props.autoPageSize],
   );
 
   /**
