+++ packages/grid/_modules_/grid/hooks/features/rowGrouping/useGridRowGrouping.tsx
@@ -22,18 +22,12 @@ import { DataGridProProcessedProps } from '../../../models/props/DataGridProProp
 import {
   filterRowTreeFromGroupingColumns,
   getRowGroupingFieldFromGroupingCriteria,
-  getColDefOverrides,
   GROUPING_COLUMNS_FEATURE_NAME,
   isGroupingColumn,
   mergeStateWithRowGroupingModel,
 } from './gridRowGroupingUtils';
-import {
-  createGroupingColDefForOneGroupingCriteria,
-  createGroupingColDefForAllGroupingCriteria,
-} from './createGroupingColDef';
 import { isDeepEqual } from '../../../utils/utils';
 import { GridPreProcessor, useGridRegisterPreProcessor } from '../../core/preProcessing';
-import { GridColumnRawLookup, GridColumnsRawState } from '../columns/gridColumnsInterfaces';
 import { useGridRegisterFilteringMethod } from '../filter/useGridRegisterFilteringMethod';
 import { GridFilteringMethod } from '../filter/gridFilterState';
 import { gridRowIdsSelector, gridRowTreeSelector } from '../rows';
@@ -41,12 +35,21 @@ import { useGridRegisterSortingMethod } from '../sorting/useGridRegisterSortingM
 import { GridSortingMethod } from '../sorting/gridSortingState';
 import { sortRowTree } from '../../../utils/tree/sortRowTree';
 import { gridFilteredDescendantCountLookupSelector } from '../filter';
-import { useGridStateInit } from '../../utils/useGridStateInit';
 import { GridRowGroupingApi, GridRowGroupingModel } from './gridRowGroupingInterfaces';
 import { useGridApiMethod } from '../../utils';
 import { gridColumnLookupSelector } from '../columns';
 import { GridRowGroupableColumnMenuItems } from '../../../components/menu/columnMenu/GridRowGroupableColumnMenuItems';
 import { GridRowGroupingColumnMenuItems } from '../../../components/menu/columnMenu/GridRowGroupingColumnMenuItems';
+import { GridStateInitializer } from '../../utils/useGridInitializeState';
+
+export const rowGroupingStateInitializer: GridStateInitializer<
+  Pick<DataGridProProcessedProps, 'rowGroupingModel' | 'initialState'>
+> = (state, props) => ({
+  ...state,
+  rowGrouping: {
+    model: props.rowGroupingModel ?? props.initialState?.rowGrouping?.model ?? [],
+  },
+});
 
 /**
  * Only available in DataGridPro
@@ -69,13 +72,6 @@ export const useGridRowGrouping = (
     | 'disableRowGrouping'
   >,
 ) => {
-  useGridStateInit(apiRef, (state) => ({
-    ...state,
-    rowGrouping: {
-      model: props.rowGroupingModel ?? props.initialState?.rowGrouping?.model ?? [],
-    },
-  }));
-
   apiRef.current.unstable_updateControlState({
     stateId: 'rowGrouping',
     propModel: props.rowGroupingModel,
@@ -214,93 +210,6 @@ export const useGridRowGrouping = (
   /**
    * PRE-PROCESSING
    */
-  const getGroupingColDefs = React.useCallback(
-    (columnsState: GridColumnsRawState) => {
-      if (props.disableRowGrouping) {
-        return [];
-      }
-
-      const groupingColDefProp = props.groupingColDef;
-
-      // We can't use `gridGroupingRowsSanitizedModelSelector` here because the new columns are not in the state yet
-      const rowGroupingModel = gridRowGroupingModelSelector(apiRef).filter(
-        (field) => !!columnsState.lookup[field],
-      );
-
-      if (rowGroupingModel.length === 0) {
-        return [];
-      }
-
-      switch (props.rowGroupingColumnMode) {
-        case 'single': {
-          return [
-            createGroupingColDefForAllGroupingCriteria({
-              apiRef,
-              rowGroupingModel,
-              colDefOverride: getColDefOverrides(groupingColDefProp, rowGroupingModel),
-              columnsLookup: columnsState.lookup,
-            }),
-          ];
-        }
-
-        case 'multiple': {
-          return rowGroupingModel.map((groupingCriteria) =>
-            createGroupingColDefForOneGroupingCriteria({
-              groupingCriteria,
-              colDefOverride: getColDefOverrides(groupingColDefProp, [groupingCriteria]),
-              groupedByColDef: columnsState.lookup[groupingCriteria],
-              columnsLookup: columnsState.lookup,
-            }),
-          );
-        }
-
-        default: {
-          return [];
-        }
-      }
-    },
-    [apiRef, props.groupingColDef, props.rowGroupingColumnMode, props.disableRowGrouping],
-  );
-
-  const updateGroupingColumn = React.useCallback(
-    (columnsState: GridColumnsRawState) => {
-      const groupingColDefs = getGroupingColDefs(columnsState);
-      let newColumnFields: string[] = [];
-      const newColumnsLookup: GridColumnRawLookup = {};
-
-      // We only keep the non-grouping columns
-      columnsState.all.forEach((field) => {
-        if (!isGroupingColumn(field)) {
-          newColumnFields.push(field);
-          newColumnsLookup[field] = columnsState.lookup[field];
-        }
-      });
-
-      // We add the grouping column
-      groupingColDefs.forEach((groupingColDef) => {
-        const matchingGroupingColDef = columnsState.lookup[groupingColDef.field];
-        if (matchingGroupingColDef) {
-          groupingColDef.width = matchingGroupingColDef.width;
-          groupingColDef.flex = matchingGroupingColDef.flex;
-        }
-
-        newColumnsLookup[groupingColDef.field] = groupingColDef;
-      });
-      const startIndex = newColumnFields[0] === '__check__' ? 1 : 0;
-      newColumnFields = [
-        ...newColumnFields.slice(0, startIndex),
-        ...groupingColDefs.map((colDef) => colDef.field),
-        ...newColumnFields.slice(startIndex),
-      ];
-
-      columnsState.all = newColumnFields;
-      columnsState.lookup = newColumnsLookup;
-
-      return columnsState;
-    },
-    [getGroupingColDefs],
-  );
-
   const addColumnMenuButtons = React.useCallback(
     (initialValue: JSX.Element[], columns: GridStateColDef) => {
       if (props.disableRowGrouping) {
@@ -352,7 +261,6 @@ export const useGridRowGrouping = (
     [apiRef],
   );
 
-  useGridRegisterPreProcessor(apiRef, 'hydrateColumns', updateGroupingColumn);
   useGridRegisterPreProcessor(apiRef, 'columnMenu', addColumnMenuButtons);
   useGridRegisterFilteringMethod(apiRef, GROUPING_COLUMNS_FEATURE_NAME, filteringMethod);
   useGridRegisterSortingMethod(apiRef, GROUPING_COLUMNS_FEATURE_NAME, sortingMethod);
