+++ packages/grid/_modules_/grid/hooks/features/rowGrouping/useGridRowGroupingPreProcessors.ts
@@ -1,0 +1,108 @@
+import * as React from 'react';
+import { GridApiRef } from '../../../models/api/gridApiRef';
+import { GridColumnsRawState, GridColumnRawLookup } from '../columns/gridColumnsInterfaces';
+import { DataGridProProcessedProps } from '../../../models/props/DataGridProProps';
+import { gridRowGroupingModelSelector } from './gridRowGroupingSelector';
+import {
+  createGroupingColDefForAllGroupingCriteria,
+  createGroupingColDefForOneGroupingCriteria,
+} from './createGroupingColDef';
+import { getColDefOverrides, isGroupingColumn } from './gridRowGroupingUtils';
+import { useGridRegisterPreProcessor } from '../../core/preProcessing/useGridRegisterPreProcessor';
+
+export const useGridRowGroupingPreProcessors = (
+  apiRef: GridApiRef,
+  props: Pick<
+    DataGridProProcessedProps,
+    'disableRowGrouping' | 'groupingColDef' | 'rowGroupingColumnMode'
+  >,
+) => {
+  const getGroupingColDefs = React.useCallback(
+    (columnsState: GridColumnsRawState) => {
+      if (props.disableRowGrouping) {
+        return [];
+      }
+
+      const groupingColDefProp = props.groupingColDef;
+
+      // We can't use `gridGroupingRowsSanitizedModelSelector` here because the new columns are not in the state yet
+      const rowGroupingModel = gridRowGroupingModelSelector(apiRef).filter(
+        (field) => !!columnsState.lookup[field],
+      );
+
+      if (rowGroupingModel.length === 0) {
+        return [];
+      }
+
+      switch (props.rowGroupingColumnMode) {
+        case 'single': {
+          return [
+            createGroupingColDefForAllGroupingCriteria({
+              apiRef,
+              rowGroupingModel,
+              colDefOverride: getColDefOverrides(groupingColDefProp, rowGroupingModel),
+              columnsLookup: columnsState.lookup,
+            }),
+          ];
+        }
+
+        case 'multiple': {
+          return rowGroupingModel.map((groupingCriteria) =>
+            createGroupingColDefForOneGroupingCriteria({
+              groupingCriteria,
+              colDefOverride: getColDefOverrides(groupingColDefProp, [groupingCriteria]),
+              groupedByColDef: columnsState.lookup[groupingCriteria],
+              columnsLookup: columnsState.lookup,
+            }),
+          );
+        }
+
+        default: {
+          return [];
+        }
+      }
+    },
+    [apiRef, props.groupingColDef, props.rowGroupingColumnMode, props.disableRowGrouping],
+  );
+
+  const updateGroupingColumn = React.useCallback(
+    (columnsState: GridColumnsRawState) => {
+      const groupingColDefs = getGroupingColDefs(columnsState);
+      let newColumnFields: string[] = [];
+      const newColumnsLookup: GridColumnRawLookup = {};
+
+      // We only keep the non-grouping columns
+      columnsState.all.forEach((field) => {
+        if (!isGroupingColumn(field)) {
+          newColumnFields.push(field);
+          newColumnsLookup[field] = columnsState.lookup[field];
+        }
+      });
+
+      // We add the grouping column
+      groupingColDefs.forEach((groupingColDef) => {
+        const matchingGroupingColDef = columnsState.lookup[groupingColDef.field];
+        if (matchingGroupingColDef) {
+          groupingColDef.width = matchingGroupingColDef.width;
+          groupingColDef.flex = matchingGroupingColDef.flex;
+        }
+
+        newColumnsLookup[groupingColDef.field] = groupingColDef;
+      });
+      const startIndex = newColumnFields[0] === '__check__' ? 1 : 0;
+      newColumnFields = [
+        ...newColumnFields.slice(0, startIndex),
+        ...groupingColDefs.map((colDef) => colDef.field),
+        ...newColumnFields.slice(startIndex),
+      ];
+
+      columnsState.all = newColumnFields;
+      columnsState.lookup = newColumnsLookup;
+
+      return columnsState;
+    },
+    [getGroupingColDefs],
+  );
+
+  useGridRegisterPreProcessor(apiRef, 'hydrateColumns', updateGroupingColumn);
+};
