+++ packages/grid/_modules_/grid/hooks/features/rows/gridRowsState.ts
@@ -1,4 +1,28 @@
-import { GridRowGroupingResult } from '../../core/rowGroupsPerProcessing';
+import type { GridRowGroupingResult, GridRowGroupParams } from '../../core/rowGroupsPerProcessing';
+import { DataGridProcessedProps } from '../../../models/props/DataGridProps';
+import { GridRowsProp } from '../../../models/gridRows';
+
+export type GridRowInternalCacheValue = Omit<GridRowGroupParams, 'previousTree'>;
+
+export interface GridRowsInternalCacheState {
+  value: GridRowInternalCacheValue;
+  /**
+   * The value of the properties used by the grouping when the internal cache was created
+   * We are storing it instead of accessing it directly when storing the cache to avoid synchronization issues
+   */
+  props: Pick<DataGridProcessedProps, 'rowCount' | 'getRowId'>;
+  /**
+   * The rows as they were the last time all the rows have been updated at once
+   * It is used to avoid processing several time the same set of rows
+   */
+  rowsBeforePartialUpdates: GridRowsProp;
+}
+
+export interface GridRowsInternalCache {
+  state: GridRowsInternalCacheState;
+  timeout: NodeJS.Timeout | null;
+  lastUpdateMs: number;
+}
 
 export interface GridRowsState extends GridRowGroupingResult {
   /**
