+++ packages/grid/_modules_/grid/hooks/features/selection/useGridSelection.ts
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { unstable_composeClasses as composeClasses } from '@mui/material';
 import { GridEvents, GridEventListener } from '../../../models/events';
 import { DataGridProcessedProps } from '../../../models/props/DataGridProps';
 import { GridApiRef } from '../../../models/api/gridApiRef';
@@ -17,28 +16,12 @@ import {
 } from './gridSelectionSelector';
 import { gridPaginatedVisibleSortedGridRowIdsSelector } from '../pagination';
 import { gridVisibleSortedRowIdsSelector } from '../filter/gridFilterSelector';
-import { GRID_CHECKBOX_SELECTION_COL_DEF, GridColDef } from '../../../models';
-import { getDataGridUtilityClass, gridClasses } from '../../../gridClasses';
+import { GRID_CHECKBOX_SELECTION_COL_DEF } from '../../../models';
+import { gridClasses } from '../../../gridClasses';
 import { useGridStateInit } from '../../utils/useGridStateInit';
-import { GridPreProcessor, useGridRegisterPreProcessor } from '../../core/preProcessing';
 import { GridCellModes } from '../../../models/gridEditRowModel';
 import { isKeyboardEvent } from '../../../utils/keyboardUtils';
 import { getCurrentPageRows } from '../../utils/useCurrentPageRows';
-
-type OwnerState = { classes: DataGridProcessedProps['classes'] };
-
-const useUtilityClasses = (ownerState: OwnerState) => {
-  const { classes } = ownerState;
-
-  return React.useMemo(() => {
-    const slots = {
-      cellCheckbox: ['cellCheckbox'],
-      columnHeaderCheckbox: ['columnHeaderCheckbox'],
-    };
-
-    return composeClasses(slots, getDataGridUtilityClass, classes);
-  }, [classes]);
-};
 
 /**
  * @requires useGridRows (state, method)
@@ -76,8 +59,6 @@ export const useGridSelection = (
 
   useGridStateInit(apiRef, (state) => ({ ...state, selection: propSelectionModel ?? [] }));
 
-  const ownerState = { classes: props.classes };
-  const classes = useUtilityClasses(ownerState);
   const lastRowToggled = React.useRef<GridRowId | null>(null);
 
   apiRef.current.unstable_updateControlState({
@@ -121,36 +102,6 @@ export const useGridSelection = (
     },
     [apiRef],
   );
-
-  /**
-   * PRE-PROCESSING
-   */
-  const updateSelectionColumn = React.useCallback<GridPreProcessor<'hydrateColumns'>>(
-    (columnsState) => {
-      const selectionColumn: GridColDef = {
-        ...GRID_CHECKBOX_SELECTION_COL_DEF,
-        cellClassName: classes.cellCheckbox,
-        headerClassName: classes.columnHeaderCheckbox,
-        headerName: apiRef.current.getLocaleText('checkboxSelectionHeaderName'),
-      };
-
-      const shouldHaveSelectionColumn = props.checkboxSelection;
-      const haveSelectionColumn = columnsState.lookup[selectionColumn.field] != null;
-
-      if (shouldHaveSelectionColumn && !haveSelectionColumn) {
-        columnsState.lookup[selectionColumn.field] = selectionColumn;
-        columnsState.all = [selectionColumn.field, ...columnsState.all];
-      } else if (!shouldHaveSelectionColumn && haveSelectionColumn) {
-        delete columnsState.lookup[selectionColumn.field];
-        columnsState.all = columnsState.all.filter((field) => field !== selectionColumn.field);
-      }
-
-      return columnsState;
-    },
-    [apiRef, classes, props.checkboxSelection],
-  );
-
-  useGridRegisterPreProcessor(apiRef, 'hydrateColumns', updateSelectionColumn);
 
   /**
    * API METHODS
