+++ packages/grid/_modules_/grid/hooks/features/sorting/useGridSorting.ts
@@ -17,7 +17,6 @@ import {
   gridSortModelSelector,
 } from './gridSortingSelector';
 import { gridRowIdsSelector, gridRowGroupingNameSelector, gridRowTreeSelector } from '../rows';
-import { useGridStateInit } from '../../utils/useGridStateInit';
 import { useFirstRender } from '../../utils/useFirstRender';
 import { GridSortingMethod, GridSortingMethodCollection } from './gridSortingState';
 import {
@@ -27,9 +26,20 @@ import {
 } from './gridSortingUtils';
 import { GridPreProcessor, useGridRegisterPreProcessor } from '../../core/preProcessing';
 import { useGridRegisterSortingMethod } from './useGridRegisterSortingMethod';
+import { GridStateInitializer } from '../../utils/useGridInitializeState';
+
+export const sortingStateInitializer: GridStateInitializer<
+  Pick<DataGridProcessedProps, 'sortModel' | 'initialState'>
+> = (state, props) => ({
+  ...state,
+  sorting: {
+    sortModel: props.sortModel ?? props.initialState?.sorting?.sortModel ?? [],
+    sortedRows: [],
+  },
+});
 
 /**
- * @requires useGridRows (state, event)
+ * @requires useGridRows (event)
  * @requires useGridColumns (event)
  */
 export const useGridSorting = (
@@ -47,14 +57,6 @@ export const useGridSorting = (
   const logger = useGridLogger(apiRef, 'useGridSorting');
   const sortingMethodCollectionRef = React.useRef<GridSortingMethodCollection>({});
   const lastSortingMethodApplied = React.useRef<GridSortingMethod | null>(null);
-
-  useGridStateInit(apiRef, (state) => ({
-    ...state,
-    sorting: {
-      sortModel: props.sortModel ?? props.initialState?.sorting?.sortModel ?? [],
-      sortedRows: [],
-    },
-  }));
 
   apiRef.current.unstable_updateControlState({
     stateId: 'sortModel',
