+++ packages/grid/_modules_/grid/hooks/features/treeData/useGridTreeData.tsx
@@ -1,23 +1,13 @@
 import * as React from 'react';
 import { GridApiRef } from '../../../models/api/gridApiRef';
 import { DataGridProProcessedProps } from '../../../models/props/DataGridProProps';
-import {
-  GRID_TREE_DATA_GROUPING_COL_DEF,
-  GRID_TREE_DATA_GROUPING_COL_DEF_FORCED_PROPERTIES,
-} from './gridTreeDataGroupColDef';
 import { useGridApiEventHandler } from '../../utils/useGridApiEventHandler';
 import { GridEventListener, GridEvents } from '../../../models/events';
-import {
-  GridColDef,
-  GridGroupingColDefOverride,
-  GridGroupingColDefOverrideParams,
-} from '../../../models';
 import { isSpaceKey } from '../../../utils/keyboardUtils';
 import { useFirstRender } from '../../utils/useFirstRender';
 import { buildRowTree, BuildRowTreeGroupingCriteria } from '../../../utils/tree/buildRowTree';
 import { GridRowGroupingPreProcessing } from '../../core/rowGroupsPerProcessing';
 import { gridFilteredDescendantCountLookupSelector } from '../filter';
-import { GridPreProcessor, useGridRegisterPreProcessor } from '../../core/preProcessing';
 import { GridFilteringMethod } from '../filter/gridFilterState';
 import { gridRowIdsSelector, gridRowTreeSelector } from '../rows';
 import { useGridRegisterFilteringMethod } from '../filter/useGridRegisterFilteringMethod';
@@ -25,9 +15,8 @@ import { useGridRegisterSortingMethod } from '../sorting/useGridRegisterSortingM
 import { GridSortingMethod } from '../sorting/gridSortingState';
 import { sortRowTree } from '../../../utils/tree/sortRowTree';
 import { filterRowTreeFromTreeData } from './gridTreeDataUtils';
-import { GridTreeDataGroupingCell } from '../../../components';
 
-const TREE_DATA_GROUPING_NAME = 'tree-data';
+export const TREE_DATA_GROUPING_NAME = 'tree-data';
 
 /**
  * Only available in DataGridPro
@@ -113,69 +102,6 @@ export const useGridTreeData = (
   /**
    * PRE-PROCESSING
    */
-  const getGroupingColDef = React.useCallback((): GridColDef => {
-    const groupingColDefProp = props.groupingColDef;
-
-    let colDefOverride: GridGroupingColDefOverride | null | undefined;
-    if (typeof groupingColDefProp === 'function') {
-      const params: GridGroupingColDefOverrideParams = {
-        groupingName: TREE_DATA_GROUPING_NAME,
-        fields: [],
-      };
-
-      colDefOverride = groupingColDefProp(params);
-    } else {
-      colDefOverride = groupingColDefProp;
-    }
-
-    const { hideDescendantCount, ...colDefOverrideProperties } = colDefOverride ?? {};
-
-    const commonProperties: Omit<GridColDef, 'field' | 'editable'> = {
-      ...GRID_TREE_DATA_GROUPING_COL_DEF,
-      renderCell: (params) => (
-        <GridTreeDataGroupingCell {...params} hideDescendantCount={hideDescendantCount} />
-      ),
-      headerName: apiRef.current.getLocaleText('treeDataGroupingHeaderName'),
-    };
-
-    return {
-      ...commonProperties,
-      ...colDefOverrideProperties,
-      ...GRID_TREE_DATA_GROUPING_COL_DEF_FORCED_PROPERTIES,
-    };
-  }, [apiRef, props.groupingColDef]);
-
-  const updateGroupingColumn = React.useCallback<GridPreProcessor<'hydrateColumns'>>(
-    (columnsState) => {
-      const groupingColDefField = GRID_TREE_DATA_GROUPING_COL_DEF_FORCED_PROPERTIES.field;
-
-      const shouldHaveGroupingColumn = props.treeData;
-      const prevGroupingColumn = columnsState.lookup[groupingColDefField];
-
-      if (shouldHaveGroupingColumn) {
-        const newGroupingColumn = getGroupingColDef();
-        if (prevGroupingColumn) {
-          newGroupingColumn.width = prevGroupingColumn.width;
-          newGroupingColumn.flex = prevGroupingColumn.flex;
-        }
-        columnsState.lookup[groupingColDefField] = newGroupingColumn;
-        if (prevGroupingColumn == null) {
-          const index = columnsState.all[0] === '__check__' ? 1 : 0;
-          columnsState.all = [
-            ...columnsState.all.slice(0, index),
-            groupingColDefField,
-            ...columnsState.all.slice(index),
-          ];
-        }
-      } else if (!shouldHaveGroupingColumn && prevGroupingColumn) {
-        delete columnsState.lookup[groupingColDefField];
-        columnsState.all = columnsState.all.filter((field) => field !== groupingColDefField);
-      }
-
-      return columnsState;
-    },
-    [props.treeData, getGroupingColDef],
-  );
 
   const filteringMethod = React.useCallback<GridFilteringMethod>(
     (params) => {
@@ -205,7 +131,6 @@ export const useGridTreeData = (
     [apiRef, props.disableChildrenSorting],
   );
 
-  useGridRegisterPreProcessor(apiRef, 'hydrateColumns', updateGroupingColumn);
   useGridRegisterFilteringMethod(apiRef, TREE_DATA_GROUPING_NAME, filteringMethod);
   useGridRegisterSortingMethod(apiRef, TREE_DATA_GROUPING_NAME, sortingMethod);
