+++ scripts/x-data-grid.exports.json
@@ -320,6 +320,7 @@
   { "name": "GridRowId", "kind": "TypeAlias" },
   { "name": "GridRowIdGetter", "kind": "TypeAlias" },
   { "name": "gridRowIdsSelector", "kind": "Variable" },
+  { "name": "GridRowInternalCacheValue", "kind": "TypeAlias" },
   { "name": "GridRowMode", "kind": "TypeAlias" },
   { "name": "GridRowModel", "kind": "TypeAlias" },
   { "name": "GridRowModelUpdate", "kind": "Interface" },
@@ -328,6 +329,8 @@
   { "name": "GridRowProps", "kind": "Interface" },
   { "name": "GridRowScrollEndParams", "kind": "Interface" },
   { "name": "GridRowSelectionCheckboxParams", "kind": "Interface" },
+  { "name": "GridRowsInternalCache", "kind": "Interface" },
+  { "name": "GridRowsInternalCacheState", "kind": "Interface" },
   { "name": "GridRowsLookup", "kind": "TypeAlias" },
   { "name": "gridRowsLookupSelector", "kind": "Variable" },
   { "name": "GridRowsMeta", "kind": "Interface" },
